Move Message.js
